+++ ALU/decimaloperations.h
@@ -5,15 +5,15 @@
 class DecimalOperations
 {
 public:
-    DecimalOperations(long op1, long op2, long *result);
+    DecimalOperations(float op1, float op2, float *result);
     ~DecimalOperations();
     void add();
     void multiply();
     void divide();
 private:
-    long op1;
-    long op2;
-    long *result;
+    float op1;
+    float op2;
+    float *result;
 };
 
 #endif // DECIMALOPERATIONS_H
